Use same error types for both drivers
