+++ src/drivers/stspin220.rs
@@ -12,7 +12,7 @@
 //! # fn main()
 //! #     -> Result<
 //! #         (),
-//! #         step_dir::drivers::stspin220::StepError<core::convert::Infallible>
+//! #         step_dir::StepError<core::convert::Infallible>
 //! #     > {
 //! #
 //! use step_dir::{
@@ -79,10 +79,10 @@
 use embedded_hal::digital::{OutputPin, PinState};
 use embedded_time::{
     duration::{Microseconds, Nanoseconds},
-    Clock, TimeError,
+    Clock,
 };
 
-use crate::{Dir, SetStepMode, Step, StepMode256};
+use crate::{Dir, ModeError, SetStepMode, Step, StepError, StepMode256};
 
 /// The STSPIN220 driver API
 ///
@@ -353,37 +353,5 @@ pub fn step_mode_to_signals(
         M64 => (High, High, Low, High),
         M128 => (High, Low, Low, Low),
         M256 => (High, High, Low, Low),
-    }
-}
-
-/// An error that can occur while setting the microstepping mode
-#[derive(Debug, Eq, PartialEq)]
-pub enum ModeError<OutputPinError> {
-    /// An error originated from using the [`OutputPin`] trait
-    OutputPin(OutputPinError),
-
-    /// An error originated from working with a timer
-    Time(TimeError),
-}
-
-impl<OutputPinError> From<TimeError> for ModeError<OutputPinError> {
-    fn from(err: TimeError) -> Self {
-        Self::Time(err)
-    }
-}
-
-/// An error that can occur while making a step
-#[derive(Debug, Eq, PartialEq)]
-pub enum StepError<OutputPinError> {
-    /// An error originated from using the [`OutputPin`] trait
-    OutputPin(OutputPinError),
-
-    /// An error originated from working with a timer
-    Time(TimeError),
-}
-
-impl<OutputPinError> From<TimeError> for StepError<OutputPinError> {
-    fn from(err: TimeError) -> Self {
-        Self::Time(err)
     }
 }
